requirements.txt
Flask
pytesseract
Pillow
openai